Recalculate the size of textboxes and mark fields correct
The size of textboxes differ greatly between gtk and qt. So the best option is to ignore the size that is in the file and calculate the size based on the text in it.
For merging of text templates there is a special syntax. so every field hast du be surrounded by ${ and }