Build instructions for MacOS
Here I'm adding the build instructions for MacOS. I also wrote a Brew Formula, though there's a possibility it gets discarded because of the lack of versioning on this project. Here's the [PR for the formula](https://github.com/Homebrew/homebrew-core/pull/39409)

It would be nice to add version tags on this repository, even though you don't feel it's all featured, so we might avoid issues with package managers, and any issues in future updates (in case you replace the continuous tag's files which would break the SHA256 of the formula).